readline function takes size: efficiency improved
readline function used to try to change the length of the requested
string everytime it was called. (this would only happen by the getline()
function if requested string is not NULL)

by this commit, now readline does not set string length; it uses
previously defined value. this is good because getline() calls will
update this previously defined value; hence, avoiding the further
malloc() calls.

also a mistake of specifying byte size to allocsize is fixed, it takes
number of characters as it should be.